run forge fmt over testdata

+++ testdata/cheats/Env.t.sol
@@ -13,19 +13,21 @@ contract EnvTest is DSTest {
         cheats.setEnv(key, val);
     }
 
-    uint constant numEnvBoolTests = 4;
+    uint256 constant numEnvBoolTests = 4;
+
     function testEnvBool() public {
         string memory key = "_foundryCheatcodeEnvBoolTestKey";
         string[numEnvBoolTests] memory values = ["true", "false", "True", "False"];
         bool[numEnvBoolTests] memory expected = [true, false, true, false];
-        for(uint i = 0; i < numEnvBoolTests; ++i){
+        for (uint256 i = 0; i < numEnvBoolTests; ++i) {
             cheats.setEnv(key, values[i]);
             bool output = cheats.envBool(key);
             require(output == expected[i], "envBool failed");
         }
     }
 
-    uint constant numEnvUintTests = 4;
+    uint256 constant numEnvUintTests = 4;
+
     function testEnvUint() public {
         string memory key = "_foundryCheatcodeEnvUintTestKey";
         string[numEnvUintTests] memory values = [
@@ -34,20 +36,17 @@ contract EnvTest is DSTest {
             "0x0000000000000000000000000000000000000000000000000000000000000000",
             "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"
         ];
-        uint256[numEnvUintTests] memory expected = [
-            type(uint256).min,
-            type(uint256).max,
-            type(uint256).min,
-            type(uint256).max
-        ];
-        for(uint i = 0; i < numEnvUintTests; ++i){
+        uint256[numEnvUintTests] memory expected =
+            [type(uint256).min, type(uint256).max, type(uint256).min, type(uint256).max];
+        for (uint256 i = 0; i < numEnvUintTests; ++i) {
             cheats.setEnv(key, values[i]);
             uint256 output = cheats.envUint(key);
             require(output == expected[i], "envUint failed");
         }
     }
 
-    uint constant numEnvIntTests = 4;
+    uint256 constant numEnvIntTests = 4;
+
     function testEnvInt() public {
         string memory key = "_foundryCheatcodeEnvIntTestKey";
         string[numEnvIntTests] memory values = [
@@ -56,89 +55,72 @@ contract EnvTest is DSTest {
             "-0x8000000000000000000000000000000000000000000000000000000000000000",
             "0x7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"
         ];
-        int256[numEnvIntTests] memory expected = [
-            type(int256).min,
-            type(int256).max,
-            type(int256).min,
-            type(int256).max
-        ];
-        for(uint i = 0; i < numEnvIntTests; ++i){
+        int256[numEnvIntTests] memory expected =
+            [type(int256).min, type(int256).max, type(int256).min, type(int256).max];
+        for (uint256 i = 0; i < numEnvIntTests; ++i) {
             cheats.setEnv(key, values[i]);
             int256 output = cheats.envInt(key);
             require(output == expected[i], "envInt failed");
         }
     }
 
-    uint constant numEnvAddressTests = 2;
+    uint256 constant numEnvAddressTests = 2;
+
     function testEnvAddress() public {
         string memory key = "_foundryCheatcodeEnvAddressTestKey";
-        string[numEnvAddressTests] memory values = [
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D",
-            "0x0000000000000000000000000000000000000000"
-        ];
-        address[numEnvAddressTests] memory expected = [
-            0x7109709ECfa91a80626fF3989D68f67F5b1DD12D,
-            0x0000000000000000000000000000000000000000
-        ];
-        for(uint i = 0; i < numEnvAddressTests; ++i){
+        string[numEnvAddressTests] memory values =
+            ["0x7109709ECfa91a80626fF3989D68f67F5b1DD12D", "0x0000000000000000000000000000000000000000"];
+        address[numEnvAddressTests] memory expected =
+            [0x7109709ECfa91a80626fF3989D68f67F5b1DD12D, 0x0000000000000000000000000000000000000000];
+        for (uint256 i = 0; i < numEnvAddressTests; ++i) {
             cheats.setEnv(key, values[i]);
             address output = cheats.envAddress(key);
             require(output == expected[i], "envAddress failed");
         }
     }
 
-    uint constant numEnvBytes32Tests = 2;
+    uint256 constant numEnvBytes32Tests = 2;
+
     function testEnvBytes32() public {
         string memory key = "_foundryCheatcodeEnvBytes32TestKey";
-        string[numEnvBytes32Tests] memory values = [
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D",
-            "0x00"
-        ];
+        string[numEnvBytes32Tests] memory values = ["0x7109709ECfa91a80626fF3989D68f67F5b1DD12D", "0x00"];
         bytes32[numEnvBytes32Tests] memory expected = [
             bytes32(0x7109709ECfa91a80626fF3989D68f67F5b1DD12D000000000000000000000000),
             bytes32(0x0000000000000000000000000000000000000000000000000000000000000000)
         ];
-        for(uint i = 0; i < numEnvBytes32Tests; ++i){
+        for (uint256 i = 0; i < numEnvBytes32Tests; ++i) {
             cheats.setEnv(key, values[i]);
             bytes32 output = cheats.envBytes32(key);
             require(output == expected[i], "envBytes32 failed");
         }
     }
 
-    uint constant numEnvStringTests = 2;
+    uint256 constant numEnvStringTests = 2;
+
     function testEnvString() public {
         string memory key = "_foundryCheatcodeEnvStringTestKey";
-        string[numEnvStringTests] memory values = [
-            "hello, world!",
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D"
-        ];
-        string[numEnvStringTests] memory expected = [
-            "hello, world!",
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D"
-        ];
-        for(uint i = 0; i < numEnvStringTests; ++i){
+        string[numEnvStringTests] memory values = ["hello, world!", "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D"];
+        string[numEnvStringTests] memory expected = ["hello, world!", "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D"];
+        for (uint256 i = 0; i < numEnvStringTests; ++i) {
             cheats.setEnv(key, values[i]);
             string memory output = cheats.envString(key);
             assertEq(output, expected[i], "envString failed");
         }
     }
 
-    uint constant numEnvBytesTests = 2;
+    uint256 constant numEnvBytesTests = 2;
+
     function testEnvBytes() public {
         string memory key = "_foundryCheatcodeEnvBytesTestKey";
-        string[numEnvBytesTests] memory values = [
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D",
-            "0x00"
-        ];
+        string[numEnvBytesTests] memory values = ["0x7109709ECfa91a80626fF3989D68f67F5b1DD12D", "0x00"];
         bytes[] memory expected = new bytes[](numEnvBytesTests);
         expected[0] = hex"7109709ECfa91a80626fF3989D68f67F5b1DD12D";
         expected[1] = hex"00";
-        for(uint i = 0; i < numEnvBytesTests; ++i){
+        for (uint256 i = 0; i < numEnvBytesTests; ++i) {
             cheats.setEnv(key, values[i]);
             bytes memory output = cheats.envBytes(key);
             require(
-                keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected[i]))),
-                "envBytes failed"
+                keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected[i]))), "envBytes failed"
             );
         }
     }
@@ -151,82 +133,53 @@ contract EnvTest is DSTest {
         cheats.setEnv(key, value);
         string memory delimiter = ",";
         bool[] memory output = cheats.envBool(key, delimiter);
-        require(
-            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))),
-            "envBoolArr failed"
-        );
+        require(keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))), "envBoolArr failed");
     }
 
     function testEnvUintArr() public {
         string memory key = "_foundryCheatcodeEnvUintArrTestKey";
-        string memory value = 
-            "0,"
-            "115792089237316195423570985008687907853269984665640564039457584007913129639935,"
+        string memory value = "0," "115792089237316195423570985008687907853269984665640564039457584007913129639935,"
             "0x0000000000000000000000000000000000000000000000000000000000000000,"
             "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF";
-        uint[4] memory expected = [
-            type(uint256).min,
-            type(uint256).max,
-            type(uint256).min,
-            type(uint256).max
-        ];
+        uint256[4] memory expected = [type(uint256).min, type(uint256).max, type(uint256).min, type(uint256).max];
 
         cheats.setEnv(key, value);
         string memory delimiter = ",";
-        uint[] memory output = cheats.envUint(key, delimiter);
-        require(
-            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))),
-            "envUintArr failed"
-        );
+        uint256[] memory output = cheats.envUint(key, delimiter);
+        require(keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))), "envUintArr failed");
     }
 
     function testEnvIntArr() public {
         string memory key = "_foundryCheatcodeEnvIntArrTestKey";
-        string memory value =
-            "-57896044618658097711785492504343953926634992332820282019728792003956564819968,"
+        string memory value = "-57896044618658097711785492504343953926634992332820282019728792003956564819968,"
             "57896044618658097711785492504343953926634992332820282019728792003956564819967,"
             "-0x8000000000000000000000000000000000000000000000000000000000000000,"
             "0x7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF";
-        int[4] memory expected = [
-            type(int256).min,
-            type(int256).max,
-            type(int256).min,
-            type(int256).max
-        ];
+        int256[4] memory expected = [type(int256).min, type(int256).max, type(int256).min, type(int256).max];
 
         cheats.setEnv(key, value);
         string memory delimiter = ",";
-        int[] memory output = cheats.envInt(key, delimiter);
-        require(
-            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))),
-            "envIntArr failed"
-        );
+        int256[] memory output = cheats.envInt(key, delimiter);
+        require(keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))), "envIntArr failed");
     }
 
     function testEnvAddressArr() public {
         string memory key = "_foundryCheatcodeEnvAddressArrTestKey";
-        string memory value =
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D,"
-            "0x0000000000000000000000000000000000000000";
-        address[2] memory expected = [
-            0x7109709ECfa91a80626fF3989D68f67F5b1DD12D,
-            0x0000000000000000000000000000000000000000
-        ];
+        string memory value = "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D," "0x0000000000000000000000000000000000000000";
+        address[2] memory expected =
+            [0x7109709ECfa91a80626fF3989D68f67F5b1DD12D, 0x0000000000000000000000000000000000000000];
 
         cheats.setEnv(key, value);
         string memory delimiter = ",";
         address[] memory output = cheats.envAddress(key, delimiter);
         require(
-            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))),
-            "envAddressArr failed"
+            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))), "envAddressArr failed"
         );
     }
 
     function testEnvBytes32Arr() public {
         string memory key = "_foundryCheatcodeEnvBytes32ArrTestKey";
-        string memory value =
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D,"
-            "0x00";
+        string memory value = "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D," "0x00";
         bytes32[2] memory expected = [
             bytes32(0x7109709ECfa91a80626fF3989D68f67F5b1DD12D000000000000000000000000),
             bytes32(0x0000000000000000000000000000000000000000000000000000000000000000)
@@ -236,25 +189,19 @@ contract EnvTest is DSTest {
         string memory delimiter = ",";
         bytes32[] memory output = cheats.envBytes32(key, delimiter);
         require(
-            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))),
-            "envBytes32Arr failed"
+            keccak256(abi.encodePacked((output))) == keccak256(abi.encodePacked((expected))), "envBytes32Arr failed"
         );
     }
 
     function testEnvStringArr() public {
         string memory key = "_foundryCheatcodeEnvStringArrTestKey";
-        string memory value =
-            "hello, world!|"
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D";
-        string[2] memory expected = [
-            "hello, world!",
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D"
-        ];
+        string memory value = "hello, world!|" "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D";
+        string[2] memory expected = ["hello, world!", "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D"];
 
         cheats.setEnv(key, value);
         string memory delimiter = "|";
         string[] memory output = cheats.envString(key, delimiter);
-        for (uint i = 0; i < expected.length; ++i) {
+        for (uint256 i = 0; i < expected.length; ++i) {
             require(
                 keccak256(abi.encodePacked((output[i]))) == keccak256(abi.encodePacked((expected[i]))),
                 "envStringArr failed"
@@ -264,9 +211,7 @@ contract EnvTest is DSTest {
 
     function testEnvBytesArr() public {
         string memory key = "_foundryCheatcodeEnvBytesArrTestKey";
-        string memory value =
-            "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D,"
-            "0x00";
+        string memory value = "0x7109709ECfa91a80626fF3989D68f67F5b1DD12D," "0x00";
         bytes[] memory expected = new bytes[](numEnvBytesTests);
         expected[0] = hex"7109709ECfa91a80626fF3989D68f67F5b1DD12D";
         expected[1] = hex"00";
@@ -274,7 +219,7 @@ contract EnvTest is DSTest {
         cheats.setEnv(key, value);
         string memory delimiter = ",";
         bytes[] memory output = cheats.envBytes(key, delimiter);
-        for (uint i = 0; i < expected.length; ++i) {
+        for (uint256 i = 0; i < expected.length; ++i) {
             require(
                 keccak256(abi.encodePacked((output[i]))) == keccak256(abi.encodePacked((expected[i]))),
                 "envBytesArr failed"
